handled - getImageURL to prefix with the right host if not already http-ized - date sort on multiple events - bug in upload images (not conextualing images path on upload) - admin rights adding/deleting boats
added admin function to send invite email for new site

+++ src/main/java/com/yeswesail/rest/UploadFiles.java
@@ -243,13 +243,17 @@ public class UploadFiles {
 		}
 		else
 		{
-			File directory = new File(destPath);
+			String contextPath = getContextPath(destPath);
+			File directory = new File(contextPath);
 	        File[] fList = directory.listFiles();
-	        for (File file : fList)
+	        if (fList != null)
 	        {
-	            if (!file.isFile() || (!file.getName().startsWith(prefix)))
-	            	continue;
-	            file.delete();
+		        for (File file : fList)
+		        {
+		            if (!file.isFile() || (!file.getName().startsWith(prefix)))
+		            	continue;
+		            file.delete();
+		        }
 	        }
 		}
 		lastIndex++;
